Stop overriding scrollbyprecisepixel granularity
More info clarified on the bug.
Web tests use by default precise scrolling granularity, which is being overridden by the web event builder when the feature is on. This is incorrect, and it makes thing like touch input tests use PBS when they shouldn't.
When doing this change, several tests had to be explicitly told to not use precise scrolling granularity when the percent based scrolling feature is on in order to actually test the feature.

Bug: 1330006
Change-Id: I93a53a0e39c1e5a1b685b9e07341f81cbb03b529
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3673276
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1016007}
NOKEYCHECK=True
GitOrigin-RevId: 479ac0ebbd9bdf334a1d516e7db0f0d757e915de