[blinkpy] Derive smoke test default from port OS version, not host
'copy-existing-baselines-internal' did not skip predecessor Mac
versions that only ran smoke tests by default. This was because the
smoke test default used the rebaselining system's OS version, which is
'linux-trusty' on the importer, not the version of the port copied to.

Fixed: 1322304
Change-Id: I1bde713b06295eccb8eec93ce9f1916803a8d8c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3715655
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1016228}
NOKEYCHECK=True
GitOrigin-RevId: 72aaf7c50015c86bc4a115ab0e4718182610133b