Speculative fix for missing treescope
The repro for this crash is not consistent, but from the stack trace
it appears the current node is not in a tree scope. Add a test for
this prior to trying to move focus to the node.

Bug: 1336825
Change-Id: Ie5535901298420eb2d199372c6f9bdb717efdec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3707944
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016340}
NOKEYCHECK=True
GitOrigin-RevId: 8da582e9036f76bce38d11da89a295b968116cd7