Add deleted constructor with double parameters for some geometry types
This is to avoid unwanted implicit conversion from float or double to
int when the constructor with int parameters is chosen.

Another way is to add full constructors with float parameters, but this
CL keeps consistent with the existing core/layout/geometry types.

Also for consistency, the original LayoutSize(float, float) constructor
is removed.

Bug: 1329021
Change-Id: I1eeccb7c7153f274a46cb09ac5c79263a7c7a9a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708348
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016511}
NOKEYCHECK=True
GitOrigin-RevId: f8562586c0166f6fc6b941b8cf5bde7a65c0ca6e