Stop inheriting the full PolicyContainer from outer parent.
This patch:
https://chromium-review.googlesource.com/c/chromium/src/+/3645368
introduced inheriting PolicyContainer from the parent for the initial
empty document in FencedFrame/GuestView/Portals.

The underlying reason was that they wanted to reuse the same
SiteInstance for the initial empty document for both the outer tree and
the inner tree, and avoid allocating a new renderer process. This way,
they can reuse the same process. Inheriting the SiteInstance without inheriting
COOP/COEP would create an inconsistency, because the second are used to compute
the first during a navigation. It caused one test to trigger a DCHECK.
https://chromium-review.googlesource.com/c/chromium/src/+/3705209
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/site_instance_impl.cc;l=611-615?q=site_instance_impl.cc&ss=chromium

This patch revert part of it, in order to inherit only COOP/COEP. In
particular, we don't want to inherit the PolicyContainer's anonymous
bit. This would violate the invariant that no top-level document can be
anonymous.

Bug: 1325733
Change-Id: I07eed2b603814f2ba604bf2843c8469d2ad2d3c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3706879
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016812}
NOKEYCHECK=True
GitOrigin-RevId: 84a8038e1176cc6ddd3ea491ba5258917a1a2b43