Changing PaintFlags get/setAlpha to use only floats
This CL is part of the effort of increasing color precision and
color gamut support in chrome.
https://docs.google.com/document/d/1y5qhBuK7sFkXt7c6vcrAJeIofj3h2UQ3btX3Hv1hrz8/edit#heading=h.69111d60ji6h
particularly
https://docs.google.com/document/d/1kvl_hwnIG3lAXcSxRndgcqbDGEvGoiD52-VwOuVfveE/edit?resourcekey=0-uvlZUrkCQ-o0H8ohYepsqA
To do so we are changing viz, cc, and other chrome parts to use
SkColor4f instead of SkColor.

PaintFlags already stores SkColor4f instead of SkColor, this CL is only
changing the getAlpha and setAlpha interface, as to make explicit from
users of the code that alpha is a float now. This CL saves some
conversions between SkColor4f and SkColor that shouldn't be necessary.

The next steps for PaintFlags would be to also remove the setColor that
expects SkColor.
The next steps for some of the callers of PaintFlags::getAlpha/setAlpha
would be to make them use the float received without having to convert
it to int.

Bug: 1308932
Change-Id: I3f0f1e9d134f08c21b5c6de1e2f7b67744ea7fea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696922
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Reviewed-by: Collin Baker <collinbaker@google.com>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016997}
NOKEYCHECK=True
GitOrigin-RevId: 3cd7b9e34437a6babeeae8ed2184cc002ac32941