Dispatch page up/down events on scroll up/down AT actions.
To achieve this, we also move the code from RenderAccessibilityImpl down
to the blink layer, in a new method called AXObject::Scroll(). The code
is run through the actions API.

To be able to test the feature, we added a way to trigger scroll up/down
actions through WebAXObjectProxy.

Finally, we expose the scroll actions in those platforms that support
it, by adding the actions to the node data.

Bug: 1099069
Change-Id: I345c902f7d92b83468d0eadf3031eab06ef0fd30
AX-Relnotes: generate page up/down keyboard events on AT scroll actions.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3583725
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Jacobo Aragunde Pérez <jaragunde@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1017124}
NOKEYCHECK=True
GitOrigin-RevId: d6a115dbdedff9ad8a233b6afd63f76c84025fde