[Web Payment] Add a flag for identity in can-make-payment event
Before this patch, accessing merchant and user identity fields in the
"canmakepayment" event would print a deprecation message in dev tools.

This patch adds an enabled-by-default flag for passing the merchant and
user identity fields in the "canmakepayment" event for a service worker
and the IS_READY_TO_PAY intent for an Android payment app.

After this patch, Chrome's existing default behavior is not changed,
but if chrome://flags/#identity-in-can-make-payment flag is disabled,
then the "canmakepayment" event and the IS_READY_TO_PAY intent do not
pass the merchant and user identity to the payment handler.

Blink-dev intent:
https://groups.google.com/a/chromium.org/g/blink-dev/c/AM2bwKxXacQ/m/2V91nC9JHQAJ

Bug: 1317941, 1290492
Change-Id: I676b567947267b439743b6ef5e1936f4f95cc319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3714562
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017464}
NOKEYCHECK=True
GitOrigin-RevId: fb386b14ddca612ee11b7f02518e26f4e2c8ff16