blink: adds lookup function for attribute names
This adds a lookup function for attribute names similar to that
of tag names, and wires it up in AtomicHTMLToken (behind a
feature). As AtomicHTMLToken ends up creating Attributes,
which needs a QualifiedName, it is worth caching these for
improved parsing performance.

BUG=1338583

Change-Id: I6a2c7c872e441a3fc20c1e4f85911059a3038d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3709664
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017494}
NOKEYCHECK=True
GitOrigin-RevId: 4bc8589a89425f024135121f19a8fba93b11d839