[@container] Support @container style() parsing
The implementation currently does not support !important inside style()
features. The spec says the syntax for a style-feature is the same as
for a declaration, which means !important should be allowed. Adding a
failing test for that.

The evaluation code is currently never matching.

Bug: 1302630
Change-Id: I23d0477ce1459201375b758fff34fe7aca230599
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3714615
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017599}
NOKEYCHECK=True
GitOrigin-RevId: dc67f73c4f2aa298707700036d540d34256b306b