[ic-unit] Add |IdeographicFullWidth|
This patch adds |FontMetrics::IdeographicFullWidth|, which
provides the value for the [`ic` unit].

The spec is not very clear as to how to compute the advance,
but further discussion[1] clarified that it should not
require shaping. This is confirmed to be in sync with the
current Gecko and WebKit implementations.

[`ic` unit]: https://drafts.csswg.org/css-values-4/#ic
[1]: https://github.com/w3c/csswg-drafts/issues/5498#issuecomment-686902802

Bug: 937101
Change-Id: Ia194c08bb93d84b6215b1c5aa49031c931182d59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3711076
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017843}
NOKEYCHECK=True
GitOrigin-RevId: 1d6f40f434ed7bed90078583695d075f6010e754