Block Badging API call from the service worker in fenced frames
This CL is a follow-up CL for crrev.com/c/3688774 , block powerful APIs
being called from the service worker in fenced frame. Badging API on the
window side is blocked after this CL crrev.com/c/3697562 , and this CL
will add the error handling when it's called from the service worker.

This CL introduces the copy of
ServiceWorkerRegistration::ancestor_frame_type to ServiceWorkerVersion
and ServiceWorkerVersionBaseInfo. The info is passed up to
BindBadgeServiceForServiceWorker(), and the frame type is validated by
BadgeManager.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/service_worker/service_worker_host.cc;l=117;drc=169c6cc102b39295a5bfe2f2a176b42b1c2fe2c4;bpv=1;bpt=1

Change-Id: I334d8ae9862af6c919faf2212c3237618a31eccd
Bug: 1276419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3715421
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018131}
NOKEYCHECK=True
GitOrigin-RevId: 1660e9397c24183cab1899da4ea60c12a037c263