Avoid crash when creating nested PrePaintTreeWalkContext.
For multicol, we create two PrePaintTreeWalkContext objects for the same
LayoutObject; once for the multicol container itself, and once for the
fragmentainer(s). The latter takes place right before walking the
children, in WalkFragmentationContextRootChildren(). At that point,
we've performed InvalidatePaint() for the object, which may actually
mark for paint invalidation. :-S (some custom scrollbar stuff; see stack
trace in bug report) That's probably not great, but the two
PrePaintTreeWalkContexts need to agree on whether they need a tree
builder context or not. So just check what the parent has, rather than
asking NeedsTreeBuilderContextUpdate() again for the same LayoutObject,
because that might return a different answer, due to the fact that we
mark for invalidation during invalidation.

Bug: 1339038
Change-Id: I468369491fe5d7b734ff777bf4752c203a48e01a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3721620
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018144}
NOKEYCHECK=True
GitOrigin-RevId: 30a14cdd8c1bedc9853ea25c45cb1bd5a6cf7a43