Colocate the CSS selector list with StyleRule.
When we allocate a StyleRule, use Oilpan's AdditionalBytes system
to get memory to store the associated CSSSelectors. This puts them
(implicitly) right after the StyleRule in memory. We don't make a
CSSSelectorList in this case, but convert directly from the
CSSSelectorVector into the right area in memory. This requires
a fair bit of finagling, since the parser functions must not return
and operate on a CSSSelectorVector instead of a CSSSelectorList.
Also, we need to be careful in the CSSOM code, because it is possible
(though hopefully rare) to replace the selector text from JavaScript
and that now requires allocating a new StyleRule.

The primary benefit is that we get them on the same cache line
(most of the time), reducing the number of cache misses when testing
a RuleData's selector from two (first on StyleRule, then the
CSSSelectors) to one. This gives a modest speedup.

Also, initial parsing/setup is sped up, since we have fewer allocations
to make.

Finally, we use less memory, since we save both on allocation overhead,
the array size typically added by operator new[], and the pointer within
the StyleRule (since the location is now implicit).

This is not the only possible way to reorganize the StyleRule, and
probably also not the optimal one. In particular, it would be
possible to move StyleRule off free pointers and instead tightly packed
into an array on RuleSet. However, this would require every StyleRule
reference to be replaced with an index into the RuleSet, which in turn
this would require refactoring the parser more, as the StyleRule objects
are typically made during parsing -- and probably entail freeing
StyleRule from StyleRuleBase somehow. Doing this without incurring more
copies and temporary memory allocation would be a challenge, so we
haven't gone this way in this patch; on the other hand, it would save
even more memory than we do now.

Style microbenchmarks (Zen 2, LTO but not PGO):

  Parse (µs)             Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                1744      1721   +1.4%  [ +1.3%,  +1.4%]
  Encyclopedia             9332      9274   +0.6%  [ +0.6%,  +0.7%]
  Extension                1673      1691   -1.0%  [ -1.1%,  -0.9%]
  News                    10465     10323   +1.4%  [ +1.3%,  +1.4%]
  Search                   6383      6282   +1.6%  [ +1.5%,  +1.7%]
  Social1                 20378     20251   +0.6%  [ +0.6%,  +0.7%]
  Social2                   793       739   +7.4%  [ +7.2%,  +7.9%]
  Sports                  70172     69814   +0.5%  [ +0.4%,  +0.6%]
  Video                   45730     45032   +1.6%  [ +1.2%,  +1.9%]
  Geometric mean                            +1.5%  [ +1.5%,  +1.6%]

  Initial style (µs)     Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce               11687     11634   +0.5%  [ +0.3%,  +0.6%]
  Encyclopedia           108918    108080   +0.8%  [ +0.5%,  +0.9%]
  Extension              232951    232525   +0.2%  [ +0.1%,  +0.3%]
  News                    43415     43134   +0.7%  [ +0.6%,  +0.7%]
  Search                   2710      2750   -1.4%  [ -2.4%,  -1.2%]
  Social1                 25067     24957   +0.4%  [ +0.3%,  +0.5%]
  Social2                  1103      1079   +2.2%  [ +2.0%,  +2.8%]
  Sports                  35880     35530   +1.0%  [ +0.9%,  +1.1%]
  Video                   47046     46634   +0.9%  [ +0.5%,  +1.3%]
  Geometric mean                            +0.6%  [ +0.5%,  +0.7%]

  Recalc style (µs)      Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce               13289     13109   +1.4%  [ +1.3%,  +1.5%]
  Encyclopedia            96950     95685   +1.3%  [ +1.2%,  +1.5%]
  Extension              226266    226207   +0.0%  [ -0.1%,  +0.1%]
  News                    31985     31701   +0.9%  [ +0.8%,  +1.0%]
  Search                    242       240   +0.8%  [ +0.4%,  +1.0%]
  Social1                 17385     17106   +1.6%  [ +1.5%,  +1.7%]
  Social2                   586       583   +0.6%  [ +0.4%,  +0.9%]
  Sports                  27760     27509   +0.9%  [ +0.8%,  +1.0%]
  Video                   27924     27382   +2.0%  [ +1.8%,  +2.1%]
  Geometric mean                            +1.1%  [ +1.0%,  +1.1%]

Memory usage is a bit trickier to measure. For some reason, our
PartitionAlloc counters do not show the reduced storage used
(we are in contact with the PartitionAlloc developers to debug this),
although the Oilpan counters do show the increase in GC memory.
Running Heaptrack and measuring peak consumed memory (which includes
both GC and PartitionAlloc memory) shows (modulo some variation between
runs):

  ECommerce      3409 ->  3397 kB  ( -12 kB)
  Encyclopedia  15520 -> 15440 kB  ( -80 kB)
  Extension     18485 -> 18484 kB  (  -1 kB)
  News          10106 -> 10041 kB  ( -65 kB)
  Search         5055 ->  4984 kB  ( -71 kB)
  Social1        8951 ->  8783 kB  (-168 kB)
  Social2        2630 ->  2618 kB  ( -12 kB)
  Sports        15326 -> 14967 kB  (-359 kB)
  Video         25384 -> 25059 kB  (-325 kB)

Change-Id: I351e5f5e14e2538d01a5c66a4c68ceefa44ff613
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3716604
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018182}
NOKEYCHECK=True
GitOrigin-RevId: 2eb4b5ae3824e75dd051c9c0dab3e9a0a1567133