[Graphics] Have canvas_resource_provider.cc scope access to SharedImage
Clients are supposed to call BeginSharedImageAccessDirectCHROMIUM to
ensure that the SharedImage's GLImage counterpart (if present) is bound
to the texture *before* they invoke operations that assume that the
image is thus bound. Currently, the GLES2 command decoders catch cases
where this hasn't been done and perform the binding on-demand. However,
we're trying to eliminate this on-demand binding code as part of the
larger effort of making GLImage an implementation detail of SharedImage
(the on-demand binding requires that the decoders access GLImage
directly).

This CL fixes an offending flow in canvas_resource_provider.cc: Its
call to GetSkiaImage() results in a FrameBufferTexture2D operation being
issued, which is one of the operations for which the decoders perform
on-demand binding. By ensuring that the call to GetSkiaImage() is
preceded by BeginSharedImageAccessDirectCHROMIUM (via
CanvasResourceProvider's helper method), this CL eliminates the usage
of on-demand binding by the decoders in this flow.

For verification of the impact of the change, see the difference in test
failures between PS3 and PS4:
- PS4 adds CHECKs in the on-demand decoder binding code. This results in
  37 test failures (note: generation of these failures together with
  additional client-side debugging code in a separate exploratory CL was
  how I originally identified this case as a problem).
- PS3 has the same CHECKs but additionally has the fix in this CL. This
  results in 20 test failures (i.e., there are test failures that are
  not due to this flow, but fixing this flow eliminated 17 of the
  overall test failures).

Bug: 1323341
Change-Id: Icf4ea6aeaed2ee6fec6896f61416e4fc35b31dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3714306
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018287}
NOKEYCHECK=True
GitOrigin-RevId: 3d7614942d7e544a11aa2c5753fd5c8898638e7f