Fixing comment to explain falining test
As per discussion in the bug, this issue seems to be related to the file
just being too big to be handled by fucshia, so this test is expected to
fail.

This CL updates the comment of the bug just to make it more clear.

Bug: 1090252
Change-Id: I6295dd218ec394b675a6bbb827fe21e407129aae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597871
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018676}
NOKEYCHECK=True
GitOrigin-RevId: ff5505f8c072fc23942b164157e6793baa55f987