blink: minor cleanup of code
HTMLDocumentParser needs to clear the token at some point (because
it's passed to the tokenizer each time). If the token was a
character token the clear happened after tree construction, where as
for all other tokens it was done before. The comment as to why
this was done no longer applies as AtomicHTMLToken makes a copy
of the data. As AtomicHTMLToken copies the data, this changes the
code to clear first (which slightly simplified the code).

BUG=none
TEST=none

Change-Id: I8a59f01caea403d0080cce281b46a60323b0947f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3724016
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018732}
NOKEYCHECK=True
GitOrigin-RevId: 51b24d5ef8beae1ae834b8533f39a6e362fcee78