Upload swift toolchain to CIPD
We want to pin developer builds to a specific (stable) version of the Swift toolchain so everyone is building with the same version by default, regardless of which version of Xcode they have installed. This uploads the current stable version (5.6.2) to third-party.

Bug: 1340010
Change-Id: Id5a01a2b8080d3bb34568223f06c1045ad9c76cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3722369
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Mike Pinkerton <pinkerton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018871}
NOKEYCHECK=True
GitOrigin-RevId: 5ee47e8688c822eba057c98998f4d5977d716cc1