[@container] Clear results according to ContainerQueryEvaluator::Change
In ContainerQueryEvaluator::ContainerChanged, we currently clear
*all* results whenever we compute any Change other than kNone.
However, for computed change=kNearestContainer it is not correct to
clear results with change=kDescendantContainers, since in that
situation we're not going to recalculate the container-query-dependent
elements within nested containers, hence all the results aren't
repopulated. This causes us to miss invalidation in some cases.

Fixed by only clearing results less than or equal to the computed
change.

Also added additional assertions and comments to the test
ContainerQueryEvaluatorTest.ContainerChanged. That change was prompted
by the slightly unclear/inaccurate comment about results being cleared
whenever ContainerChanged is called. It's better to be explicit
about what we expect the results to be.

Fixed: 1340351
Change-Id: I64467104b6b0eacef5100d67f6b8f8287a57d0a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3732493
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019074}
NOKEYCHECK=True
GitOrigin-RevId: f453fc772adb835985857dafb0442afff314e8f7