Move Widget scheduling information into WidgetScheduler interface.
The MainThreadSchedulerImpl still implements the logic but it is
no longer exposed on the public WebThreadScheduler interface and
exposed on the internal blink WidgetScheduler interface.

This required some reworking how WebPagePopupImpl creates the Page.
Since Page owns the PageScheduler the page needs to be created before
the WidgetBase is created.

BUG=1097816

Change-Id: I4b8b37dcf19448fe34d4cae22ae6cc648d02af9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696543
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Carlos Caballero Grolimund <carlscab@google.com>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019161}
NOKEYCHECK=True
GitOrigin-RevId: 0eac8467e38d81d32283fb77a36638ff52d63546