[mathml] Move math@display mapping to the UA stylesheet
Currently the mapping for math-style property is performed in
MathMLElement::CollectStyleForPresentationAttribute. The fact that
displaystyle="inline" wins over display="block" (as expected by MathML3)
is only due to the fact that Element::CreatePresentationAttributeStyle()
handles attributes in alphabetical order, and that "display" turns out
to be before "displaystyle". This CL instead handles the math@display
attribute in the UA stylesheet, as it is described in the MathML Core
specification [1]. Priority is now guarantee by the fact that
presentational hints for displaystyle are author-level [2] and apply
after user-agent-level rules.

Additionally, text-align: center is removed since it does not apply to
block-level elements and centering is actually handled in
NGMathRowLayoutAlgorithm, based on the display value of the <math> tag.

This CL makes Chromium's UA sheet closer to the spec's one. Behavior
is unchanged and is already covered by existing WPT tests.

[1] https://w3c.github.io/mathml-core/#user-agent-stylesheet
[2] https://html.spec.whatwg.org/#the-css-user-agent-style-sheet-and-presentational-hints

Bug: 6606
Change-Id: I9a41910a4dcd610bac78d77fb7f6ba35fc4f467a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3726068
Reviewed-by: Manuel Rego <rego@igalia.com>
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1019180}
NOKEYCHECK=True
GitOrigin-RevId: 6370b4e683cff4f159e590e86351aa199333cfb7