Fix DCHECK failure in ComputeNGCaretPositionAfterInline with <area>
https://crrev.com/c/3620254 enabled carets to be placed before or
after an inline element, and as part of its implementation, the
functions `ComputeNGCaretPositionAfterInline` and
`ComputeNGCaretPositionBeforeInline` were added to
ng_caret_position.cc. These functions, as part of their behavior,
move an `NGInlineCursor` to the fragments associated with a
`LayoutInline`, even if it is a culled inline, and DCHECK that the
resulting cursor is not null.

However, the HTML `<area>` element appears to be a culled inline with
no descendants, and therefore there are no fragments associated with
its `LayoutInline`. As such, when trying to place the caret before or
after an `<area>` element, the DCHECK fails, and if DCHECKs are
disabled, a null dereference is produced shortly thereafter.

This patch comments out the DCHECK and returns a null
`NGCaretPosition` if the cursor happens to be null.

Bug: 1340236
Change-Id: I08ebe366482e2d52115fdb1230b591b87d3f1b1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735048
Commit-Queue: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019244}
NOKEYCHECK=True
GitOrigin-RevId: 00c5de7ff3cf051eaeab4887a39f2d3a1e997233