Make ScrollAnchorTest work with scroll unification.
The AnchorWhileDraggingScrollbar test case injects scroll gestures.
Since crrev.com/1015026, this is supported with unification, but it
requires a WebViewHelper, so it cannot be done in a RenderingTest.
Therefore, we also convert ScrollAnchorTest to be a SimTest (which
addresses a longstanding TODO).

Also make scroll deltas larger than kScrollbarDragAccuracy, to avoid
false passes.

Bug: 1322078
Change-Id: I7e4561d40353671722231983cb150606f3f389fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3732551
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019645}
NOKEYCHECK=True
GitOrigin-RevId: 679a313178521c4c3bfa690667c52ad4985aa218