Add requestPermission() to DeviceOrientationEvent and DeviceMotionEvent.
Currently implemented behind the "DeviceOrientationRequestPermission"
runtime flag.

The API is kind of a stub at the moment: we do not prompt for permission
and merely resolve the returned promises to either "granted" or "denied"
based on the current Motion Sensor permissions.

The idea is to let developers start playing with the API while also
gathering user counter information so that we can later decide when to
switch to prompting for permission, preferably by default. Right now, we
count calls to requestPermission() as well as usages of
ondevice{motion,orientation} that happen without calling
requestPermission() first.

Intent to prototype:
https://groups.google.com/a/chromium.org/g/blink-dev/c/j7sGtCBGaxA

Bug: 947112
Change-Id: I2e0209755402b288c58b039aa800ecccf64bf8a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3714175
Commit-Queue: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019740}
NOKEYCHECK=True
GitOrigin-RevId: dba32ba31fd509b9c7afd56a1ecdb6a8525ea445