Use GetPaintRecord instead of doing a new PaintPreview in Phishing Detection
Client Side Phishing Detection currently does a new PaintPreview on
page load to get the input for the model. This does a lot of layout
work that is duplicated with rendering/painting. This CL uses the
existing cache of painting results held in the LocalFrameView instead
of redoing the work of computing it.

Bug: 1338944
Change-Id: Ie2a9ea7cac212e2be54940ef8a551830c3319d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3727009
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Rohit Bhatia <bhatiarohit@google.com>
Cr-Commit-Position: refs/heads/main@{#1019799}
NOKEYCHECK=True
GitOrigin-RevId: 2b60b22f1048edcae86c00db0a3d8f9f2d8f9cb6