Reduce timing dependency in document-timeline-animation.html
The original test had a short duration, but ran to completion normally.
Short duration animations can flake due to intermittent slowdowns on
test machines. Instead, we use a long duration animation, latch on to
the start of the animation and then programmatically advance to the end
of the animation.  This test completes faster and should be more
reliable while testing the same functionality (an animation running on
the compositor can have a document timeline other than the default one).

Bug: 1339491
Change-Id: I059e6f1d7f1f365f32b1d638149de65a9d837533
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735305
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019824}
NOKEYCHECK=True
GitOrigin-RevId: f62d4e898dbb1cf44b96669bbded984a8e2bde8b