[webcodecs] Add dequeue event to encoder interfaces
Implementation is behind flag:
--enable-blink-features=WebCodecsDequeueEvent

Spec for this is still WIP: https://github.com/w3c/webcodecs/issues/280
Let's get it implemented ASAP so folks can test it and verify it meets
their needs.

The interesting part of this CL is how events are coalesced to avoid
spamming folks. The WPT tests demonstrate this best. The intent is that
we should only fire the event 1 time for every batch of decrements to
the counter. Here a "batch" is defined as any decrements that take
place between the user scripts having a chance to run. For ex, if you
queue two encodes() and we immediately process those, you will find the
encodeQueueSize = 0 when the first dequeue event fires. It would be
spammy to fire a second event (one for each of the encodes) since the
queue size remains zero.

I plan to specify the above behavior after getting some feedback from
owners/partners.

Bug: 1341116

Change-Id: I2672ae96aa03b889821996ee69c2ed427c2a0f81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3646113
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Mike Taylor <miketaylr@chromium.org>
Auto-Submit: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020222}
NOKEYCHECK=True
GitOrigin-RevId: 21d8c4a981e8523efaae59a7ea96c20d8ef84174