[MirrorSync] Wire up the manage-mirrorsync dialog to Files app
The chrome://manage-mirrorsync dialog is where to select folders to sync
and view useful information about the mirror subsystem. This wires it up
to the Files gear menu under the name "Manage synced folders". This
functionality is still behind a flag, so the UX hasn't completely been
decided.

This removes some method overrides that I forgot to remove before and a
followup CL will add an integration test to make sure this item doesn't
show when the flag is disabled. All the plumbing for the flag in the
browser test framework was blowing this CL up too large.

Bug: b:237066325
Test: browser_tests --gtest_filter=*showManageMirrorSync*
Change-Id: I4cad2fa91f76fcb4783c7978ffccaec65ddd5798
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3740039
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020705}
NOKEYCHECK=True
GitOrigin-RevId: 66cf36f6e8ca320e86c17882563e5bc4058517b3