Define RTCPeerConnection's mediaConstraints in WebIDL instead.
With this CL we no longer share the same parsing logic between the PC
and getUserMedia and it is a lot more clear which constraints are
applicable to RTCPeerConnection. This will also make it very clean to
delete (or revert the delete) of individual deprecated constraints in
the future, such as removing googIPv6 in M108 to see if it "sticks".

WebIDL parsing logic automatically ignores any unrecognized constraints
which as of https://crbug.com/1328047 is the same behavior as today's
manual parsing logic so from an API usage point of view this should be
a pure refactoring CL.

If we decide to replace getUserMedia's mediaConstraints parsing by
WebIDL as well that should be a separate GUM-specific dictionary.

fast/peerconnection/RTCPeerConnection.html is updated not to make use
of test-only constraints. For the specifics, I manually verified using
https://jsfiddle.net/04q1fmwb/1/ and printfs that constraints were
passed down successfully as well as in chrome://webrtc-internals/.

Bug: chromium:1341687
Change-Id: Iab6c3a5453bca0dfb866f526f6d57810f375c9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3743775
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020912}
NOKEYCHECK=True
GitOrigin-RevId: 0f7945ddb5714560e6c202573a02e7027a124ec0