Fix a WebGPU resize crash
Fixes a race-condition crash introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/3724099

Depending on the structure of the page OnTextureTransferred was
sometimes being called after a resize but before a new texture had
been retrieved, causing the `DCHECK(texture_)` call to fail. This
was because the patch attempted to improve on the old behavior of
recreating the swap chain on every resize.

This patch fixes it by temporarily going back to the pattern of
recreating the swap buffers on resize until a better mechanism for
invalidating it's texture contents can be put in place. TODOs
explaining what needs to be done next have been added.

Bug: dawn:1470
Change-Id: Iad39ebdee8df2a72fab5f7a7ccafa66781ef8860
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3742045
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020934}
NOKEYCHECK=True
GitOrigin-RevId: b3a1dedfc9822418b57631f9166877d22c89f4fa