Simplify GoogMediaConstraints WebIDL not to support exact/ideal.
When migrating from the old manual parsing code to WebIDL, care was
taken to ensure we supported exact/ideal. I assumed this was supported
because chrome://webrtc-internals/ previously presented these using the
"exact" syntax. (Now we just present it as googFoo: true/false.)

But as it turns out, the old parser code never supported exact/ideal!
It just looked at if the parameter was to-string-equal-to "true" or
not[1]. Meaning both googIPv6:{exact:true}} and googIPv6:{exact:false}}
were interpreted as false.

This CL makes the WebIDL less sophisticated, simplifying things to
interpret everything as a boolean. The WebIDL "to boolean" logic
interprets the presence of a dictionary as the value true, so in this
case both googIPv6:{exact:true}} and googIPv6:{exact:false}} are
interpreted as true. It's not the exact same behavior, but it's more
consistent with other boolean parameters and if we never supported
exact/ideal in the first place there shouldn't be any dependency on
that anyway. So let's keep things simple.

Using https://jsfiddle.net/nk5zowxe/ I manually tested both the
behavior or Chrome M103 Stable and this CL for comparison.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/mediastream/media_constraints_impl.cc;l=280;drc=66fb5c89fa76a7b89983dd6c08e7f67aed1d0999

Bug: chromium:1341687
Change-Id: I8e76206f7939be6d42085e0efa95e462de463ec1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3745879
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1021106}
NOKEYCHECK=True
GitOrigin-RevId: 394f11d33ba872bb4036af834d3b379842e2ee71